+++ jobbergate-api/tests/apps/test_schemas.py
@@ -1,7 +1,10 @@
+from datetime import datetime, timedelta, timezone
+
+import pendulum
 import pytest
 from pydantic import ValidationError
 
-from jobbergate_api.apps.schemas import BaseModel, LengthLimitedStr
+from jobbergate_api.apps.schemas import BaseModel, LengthLimitedStr, PydanticDateTime
 
 
 class TestLengthLimitedStr:
@@ -31,3 +34,27 @@ class TestLengthLimitedStr:
         """
         with pytest.raises(ValidationError):
             self.DummySchema(name="a" * 256)
+
+
+class TestPydanticDateTime:
+
+    class DateTimeSchema(BaseModel):
+        timestamp: PydanticDateTime
+
+    def test_valid__with_pendulum_DateTime(self):
+        expected_timestamp = pendulum.parse("2024-05-31 10:21:00")
+        instance = self.DateTimeSchema(timestamp=pendulum.parse("2024-05-31 10:21:00"))
+        assert isinstance(instance.timestamp, pendulum.DateTime)
+        assert instance.timestamp == expected_timestamp
+
+    def test_valid__with_standard_datetime(self):
+        expected_timestamp = pendulum.parse("2024-05-31 10:21:00")
+        instance = self.DateTimeSchema(timestamp=datetime(2024, 5, 31, 10, 21, 0))
+        assert isinstance(instance.timestamp, pendulum.DateTime)
+        assert instance.timestamp == expected_timestamp
+
+    def test_valid__with_string_timestamp(self):
+        expected_timestamp = pendulum.parse("2024-05-31 10:21:00")
+        instance = self.DateTimeSchema(timestamp="2024-05-31 10:21:00")
+        assert isinstance(instance.timestamp, pendulum.DateTime)
+        assert instance.timestamp == expected_timestamp